Add -m option to wipe to set minimum content area for wipe to proceed
If content is very light or sparse it may be better to not wipe at all
than wipe almost all of the content leaving a small strip. This is done
now by aborting the wipe if the detected content takes up less than the
minimum % of the page (default is 30%).