Fix race condition that could cause errors to be silently discarded
This was a nasty one. By closing the up channel, the up() function
would finish and send to the done channel. This means that the select
between err and done would be random as to which was picked, whereas
of course if there has been an error that path must be taken.