Some changes to ensure the pipeline works correctly on Windows
There were a couple of places where a file was uploaded while still open,
which resulted in an attempt to remove it, which causes an error from
Windows.

The allOCRed function also included an assumption that the path separator
would be a /, which is always correct for AWS, and correct for local on
Linux and OSX, but not for local Windows. Fixed by leaving the separator
well alone.

Also, the local connection was not stripping leading \, like it did /,
which caused an issue with Windows local.

Windows local is now tested and working, at least through wine.