Don't upload binarised pdf twice needlessly
This can also result in the file being uploaded twice simultaneously,
as up() is running in a separate goroutine. This can cause failures
on Windows as the file is attempted to be removed by one upload
process while being open to upload by the other process. Probably it
could also fail if the process completed by one (so the file was
deleted) before being started by the other.