Allow completely non-embedded builds
This enables installs straight from 'go install' or 'fyne install'. It
also means warning if a system getgbook isn't found, and erroring if
tesseract isn't found (as was done already). The location of getgbook
can therefore now be specified on the command line.

Embedded builds are enabled with the -tags embed flag, which the makefile
sets for all builds.