Don't disable the log area text widget so the text is readable
Sadly making a custom theme to make the DisabledColor more readable
also affects the disabled buttons, which makes it unclear which are
disabled. So instead just enable the log area - it's not ideal that
it can be typed in, but it's alright, and it needs to be an entry
widget so the text in it can be easily copied.